passwords: support listing all passwords
