+++ internal/cmd/password/create_test.go
@@ -1,4 +1,4 @@
-package passwords
+package password
 
 import (
 	"bytes"
@@ -14,7 +14,7 @@ import (
 	qt "github.com/frankban/quicktest"
 )
 
-func TestPassword_DeleteCmd(t *testing.T) {
+func TestPassword_CreateCmd(t *testing.T) {
 	c := qt.New(t)
 
 	var buf bytes.Buffer
@@ -25,16 +25,17 @@ func TestPassword_DeleteCmd(t *testing.T) {
 	org := "planetscale"
 	db := "planetscale"
 	branch := "development"
-	password := "mypassword"
+	name := "production-password"
+	res := &ps.DatabaseBranchPassword{Name: "foo"}
 
 	svc := &mock.PasswordsService{
-		DeleteFn: func(ctx context.Context, req *ps.DeleteDatabaseBranchPasswordRequest) error {
+		CreateFn: func(ctx context.Context, req *ps.DatabaseBranchPasswordRequest) (*ps.DatabaseBranchPassword, error) {
 			c.Assert(req.Organization, qt.Equals, org)
 			c.Assert(req.Database, qt.Equals, db)
 			c.Assert(req.Branch, qt.Equals, branch)
-			c.Assert(req.PasswordId, qt.Equals, password)
+			c.Assert(req.DisplayName, qt.Equals, name)
 
-			return nil
+			return res, nil
 		},
 	}
 
@@ -51,17 +52,10 @@ func TestPassword_DeleteCmd(t *testing.T) {
 		},
 	}
 
-	cmd := DeleteCmd(ch)
-	cmd.SetArgs([]string{db, branch, password, "--force"})
+	cmd := CreateCmd(ch)
+	cmd.SetArgs([]string{db, branch, name})
 	err := cmd.Execute()
 
 	c.Assert(err, qt.IsNil)
-	c.Assert(svc.DeleteFnInvoked, qt.IsTrue)
-
-	res := map[string]string{
-		"result":   "password deleted",
-		"password": password,
-		"branch":   branch,
-	}
-	c.Assert(buf.String(), qt.JSONEquals, res)
+	c.Assert(svc.CreateFnInvoked, qt.IsTrue)
 }
